formatting

+++ api/index.js
@@ -74,7 +74,7 @@ global.fetch = require("node-fetch");
 
     // kick off periodic processes to run every
     log.info(
-	`Running periodic processes every ${configuration.api.periodicProcessInterval / 60} minutes`
+        `Running periodic processes every ${configuration.api.periodicProcessInterval / 60} minutes`
     );
     setInterval(runPeriodicProcesses, configuration.api.periodicProcessInterval * 1000);
 
@@ -95,16 +95,16 @@ async function runPeriodicProcesses() {
 
 async function getTypeDefinitions({ configuration }) {
     if (!(await pathExists("/srv/type-definitions.json")) && configuration.api?.typeDefinitions) {
-	let response = await fetch(configuration.api?.typeDefinitions);
-	const typeDefinitions = await response.json();
-	await writeJson("/srv/type-definitions.json", typeDefinitions);
+        let response = await fetch(configuration.api?.typeDefinitions);
+        const typeDefinitions = await response.json();
+        await writeJson("/srv/type-definitions.json", typeDefinitions);
     }
     if (
-	!(await pathExists("/srv/type-definitions-lookup.json")) &&
-	configuration.api?.typeDefinitionsLookup
+        !(await pathExists("/srv/type-definitions-lookup.json")) &&
+        configuration.api?.typeDefinitionsLookup
     ) {
-	let response = await fetch(configuration.api?.typeDefinitionsLookup);
-	const typeDefinitionsLookup = await response.json();
-	await writeJson("/srv/type-definitions-lookup.json", typeDefinitionsLookup);
+        let response = await fetch(configuration.api?.typeDefinitionsLookup);
+        const typeDefinitionsLookup = await response.json();
+        await writeJson("/srv/type-definitions-lookup.json", typeDefinitionsLookup);
     }
 }
